Fix 713090 (search box in chrome)
Something about the useragent stylesheet is different in the mac version
versus other versions of chrome. In the mac version, it latches on to
type="search" argument and makes it look sexy. Linux and mac don't do
that, in fact they mess up the search box to look dumb. It seemed to be
easier to just remove it.