[fix bug 750316] Make IRC handle editable again.
Make the IRC handle editable again and show it in the UI. Deemphasize
username in favor of IRC handle (username really only shows up in the
URL).

There are probably some follow-up bugs we need to file, especially
around what search results look like and possibly around the behavior,
but this is a good start.