[fix bug 763509] Add location to profiles.
Users can add country/region, state/province, and city to their profile.
The data isn't displayed on the profile anywhere--let's do that in a
follow-up: this is the important part for right now.

You can't specify a specific region without specifying a more general
one. For example, you can't list a state without a country, or a city
without a state.

Includes tests.