Bug 929718 - Fix duplicate countries and cities in API
Because the custom manager that implements field-specific privacy
requires us to include privacy_country or privacy_city in the values()
part of the query, there's not a good way to fix this in the
query itself.

Added some code that fixes up the result list after the query
to collapse the duplicates down again and add up the population.