Bug 877315 - count vouched members in admin
In the admin list for groups, add a column to display the number of
vouched members, similar to the current column that counts all
members.

This implementation relies on the database implementing Booleans
as integers, so we can Sum() them. It works for MySQL, which is
how Mozillians is currently deployed.

A test is added that will fail if we ever try to use Mozillians
with a database where this doesn't work, which will alert us to
find another approach.