Bug 938184 - Don't allow creating groups while editing user profile
Part of the changes for curated groups is that groups will only be
creatable from the groups list page, so only allow people editing
their profiles to select groups that already exist.

The tag-it library that was being used to implement the groups
selection could not easily be configured or modified to prevent
users from selecting non-existent items, so the groups field
was changed to use django-selectable. Styling was added to make
the django-selectable field look like the current tag-it fields,
which are still being used for e.g. skills and languages.