Bug 950242 - Admin management of group membership
Allow Django admins to manage group members as inlines on the group's
admin page.

This will work fine for reasonable-sized groups, but will get unwieldy
when groups get really large.  We might be better off managing group
membership from a user's page, since a user is unlikely to belong to
so many groups that that would get too big. Unfortunately, Django currently
doesn't support nested inlines in the admin, which would be required
because we already implement the user admin by making the userprofile
an inline on the user page, and the membership would have to be an
inline on the userprofile.

A better way for admins to manage memberships might be directly in
the GroupMembership model. To make that easier, I've added a search
field to that admin changelist page that searches on a bunch of fields
in the groups and users associated with the membership records. It
shouldn't be hard to find a user's memberships or membership in a
group, or just add a new membership.

Fixes bug 950242