Fix user migration 50
This migration didn't have an up-to-date copy of the db schema
in it, so when you try to create a new migration, the new migration
tries to re-do a bunch of already migrated changes that from
migration 50, don't look as if they've been updated.

Update the schema copy in this migration to match what the database
should be at this point. This should not break anything currently,
just make the next migration we create correct.