[fix bug 974011] Fix logout button in Admin Panel.
Because we're using BrowserID we need to redirect users our custom
logout view to properly sign them website and delete the BrowserID
session.

At the same time this fix solves the 500 error that was raised when a
user clicked the logout button from the admin panel, which was due to
'registration' app not being excluded from Jingo.