[bug 1048940] Remove basket_token from stage dump.
Stage and Prod share the same Basket and ET setup which can result in
updating the email subscription preferences of users while testing on
stage.

We remove the basket tokens from stage dumps which, in combination with
the email address change of all profiles, will result in users getting
subscribed as new users in Basket/ET and therefore overcome this issue.

In the future we will have a staging Basket/ET setup connected with
mozillians stage and this change will be still needed, since the Prod
tokens will be no longer valid.